hw4/tests/requirements.txt
pytest==7.4.3
allure-pytest==2.13.2